+++ tests/__init__.py
@@ -20,6 +20,8 @@ import platform
 import os
 from concurrent.futures import ThreadPoolExecutor
 
+from cassandra import DependencyException
+
 log = logging.getLogger()
 
 def is_eventlet_monkey_patched():
@@ -27,10 +29,10 @@ def is_eventlet_monkey_patched():
         return False
     try:
         import eventlet.patcher
+        return eventlet.patcher.is_monkey_patched('socket')
+    # Yet another case related to PYTHON-1364
     except AttributeError:
         return False
-    return eventlet.patcher.is_monkey_patched('socket')
-
 
 def is_gevent_monkey_patched():
     if 'gevent.monkey' not in sys.modules:
@@ -88,17 +90,18 @@ elif "twisted" in EVENT_LOOP_MANAGER:
 elif "asyncio" in EVENT_LOOP_MANAGER:
     from cassandra.io.asyncioreactor import AsyncioConnection
     connection_class = AsyncioConnection
-
 else:
+    log.debug("Using default event loop (libev)")
     try:
         from cassandra.io.libevreactor import LibevConnection
         connection_class = LibevConnection
-    except ImportError as e:
+    except DependencyException as e:
         log.debug('Could not import LibevConnection, '
                   'using connection_class=None; '
                   'failed with error:\n {}'.format(
                       repr(e)
                   ))
+        log.debug("Will attempt to set connection class at cluster initialization")
         connection_class = None